can now disable load balancing and adaptivity, and it should have been measured with size_t from the beginning

+++ include/meshes/Mesh2D_t.cuh
@@ -19,7 +19,7 @@
 namespace SEM { namespace Device { namespace Meshes {
     class Mesh2D_t {
         public:
-            Mesh2D_t(std::filesystem::path filename, int initial_N, int maximum_N, size_t n_interpolation_points, int max_split_level, int adaptivity_interval, int load_balancing_interval, deviceFloat tolerance_min, deviceFloat tolerance_max, const SEM::Device::Entities::device_vector<deviceFloat>& polynomial_nodes, const cudaStream_t &stream);
+            Mesh2D_t(std::filesystem::path filename, int initial_N, int maximum_N, size_t n_interpolation_points, int max_split_level, size_t adaptivity_interval, size_t load_balancing_interval, deviceFloat tolerance_min, deviceFloat tolerance_max, const SEM::Device::Entities::device_vector<deviceFloat>& polynomial_nodes, const cudaStream_t &stream);
 
             // Geometry
             SEM::Device::Entities::device_vector<SEM::Device::Entities::Vec2<deviceFloat>> nodes_;
@@ -128,8 +128,8 @@ namespace SEM { namespace Device { namespace Meshes {
             int maximum_N_;
             size_t n_interpolation_points_;
             int max_split_level_;
-            int adaptivity_interval_;
-            int load_balancing_interval_;
+            size_t adaptivity_interval_;
+            size_t load_balancing_interval_;
             deviceFloat tolerance_min_;
             deviceFloat tolerance_max_;
